+++ react-app/src/pages/CreateFrom.jsx
@@ -1,0 +1,8 @@
+import FormComponent from "@/components/form-components/create-update-form";
+import React from "react";
+
+const CreateFrom = () => {
+  return <FormComponent />;
+};
+
+export default CreateFrom;
